completed part4 - ex 4.15 to 4.22, left out testing (4.23)

+++ part4/blog-list/controllers/blogs.js
@@ -1,24 +1,54 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-    const blog = await Blog.find({})
+    const blog = await Blog.find({}).populate('user', {username: 1, name: 1, id: 1})
     response.json(blog)
   })
   
-blogsRouter.post('/', (request, response) => {
-    const blog = new Blog(request.body)
+blogsRouter.post('/', async (request, response) => {
 
-    const savedBlog = blog.save()
+    const body = request.body
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    console.log(decodedToken)
+
+    if (!decodedToken.id) return response.status(400).json({error: "Invalid token"})
+
+    const user = await User.findById(decodedToken.id)
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes,
+      user: user._id
+    })
+    
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
-  try {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response, next) => {
+
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) return response.status(400).json({error: "Invalid Blog"})
+  else if (!user) return response.status(400).json({error: "Invalid User or Token"})
+
+  if (blog.user.toString() === user.id) {
     await Blog.findByIdAndRemove(request.params.id)
+    const newBlogs = user.blogs.filter(blog => blog.toString() !==  request.params.id)
+    user.blogs = newBlogs
+    await user.save()
     response.status(204).end()
-  } catch(exception) {
-    next(exception)
+  } else {
+    response.status(400).json({error: "blog post does not belong to user"})
   }
 })
